4.21

+++ controllers/blogs.js
@@ -1,24 +1,57 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 ///api/blogs
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
+    .populate('user',{ username:1, name:1, id:1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-  try{
-    const blog = new Blog(request.body)
-    const result = await blog.save()
-    response.status(201).json(result)
-  } catch(exception){
+  const body = request.body
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    console.log('decodedToken', decodedToken)
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({
+      url: body.url,
+      title: body.title,
+      author: body.author,
+      likes: body.likes,
+      user: user._id
+    })
+
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.json(savedBlog.toJSON())
+  }
+  catch(exception) {
     next(exception)
   }
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog && blog.user.toString() !== decodedToken.id.toString()) {
+      return response.status(401).end({ error: 'only owner user can delete blog' })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch(exception) {
